+++ recommender.py
@@ -28,7 +28,7 @@ COSINE_SCORE_THRESHOLD = 0.6
 # Initialize Vertex AI & Gemini
 #-----------------------------------
 PROJECT_ID = os.environ.get('MY_PROJECT_ID')  # @param {type:"string"}
-LOCATION = "northamerica-northeast1"  # @param {type:"string"}
+LOCATION = "us-central1"  # @param {type:"string"}
 
 # if not running on colab, try to get the PROJECT_ID automatically
 if "google.colab" not in sys.modules:
@@ -47,8 +47,8 @@ vertexai.init(project=PROJECT_ID, location=LOCATION)
 #multimodal_model = GenerativeModel("gemini-1.5-pro-002")
 #multimodal_model = GenerativeModel("gemini-1.5-flash-002")
 multimodal_model = GenerativeModel(
-        #"gemini-1.5-flash-002",
-        "gemini-1.5-pro-002",
+        "gemini-2.0-flash-001",
+        #"gemini-1.5-pro-002",
         system_instruction=[
             "You are a fashion stylist.",
             "Your mission is to describe the clothing you see.",
@@ -258,6 +258,7 @@ def get_reference_image_description(image_filename: str) -> list:
   )
 
   output_description_text = response.text.split('\n\n')
+  print(f"DEBUG DESCRIPTION: {output_description_text}")
   return output_description_text
 
 
@@ -272,15 +273,13 @@ from vertexai.vision_models import Image as vision_model_Image
 from vertexai.vision_models import MultiModalEmbeddingModel
 
 # for embedding
-#text_embedding_model = TextEmbeddingModel.from_pretrained("textembedding-gecko@003")
 text_embedding_model = TextEmbeddingModel.from_pretrained("text-embedding-005")
 
 
 # CSV more precise than JSON
 # skipping first column as that's an additional column number
 # GOTCHA: the column in the CSV that gets read in is read as a string rather than a list of vectors :(
-#image_metadata_df_csv = pd.read_csv("mywardrobe_1-0-pro-vision.csv",converters={"image_description_text_embedding": lambda x: x.strip("[]").split(", ")})
-image_metadata_df_csv = pd.read_csv("mywardrobe_1-5-pro.csv",converters={"image_description_text_embedding": lambda x: x.strip("[]").split(", ")})
+image_metadata_df_csv = pd.read_csv("mywardrobe_2-0-flash.csv",converters={"image_description_text_embedding": lambda x: x.strip("[]").split(", ")})
 print('=== FINDING BEST MATCHES... ===')
 
 
@@ -304,7 +303,8 @@ while retry_count < 5:
   # filter out responses that have more than 1 clothing type listed
   for query in queries:
     num_clothing_types=any_list_element_in_string(clothing_list, query)
-    if num_clothing_types > 1:
+    print(f"DEBUG NUM_CLOTHING_TYPES: {num_clothing_types}")
+    if num_clothing_types == 0:
       print("INFO: ", num_clothing_types, query)
       queries.remove(query)